+++ src/Pkg.jl
@@ -625,6 +625,7 @@ Print out the status of the project/manifest.
 Packages marked with `⌃` have new versions that can be installed, e.g. via [`Pkg.update`](@ref).
 Those marked with `⌅` have new versions available, but cannot be installed due to compatibility conflicts with other packages. To see why, set the
 keyword argument `outdated=true`.
+Packages marked with `○` are not precompiled yet.
 Packages marked with `[yanked]` are yanked versions that should be updated or replaced as they may contain bugs or security vulnerabilities.
 
 Setting `outdated=true` will only show packages that are not on the latest version,